(1248) Split up highCovidCommentModel
